REF: share Index arithmetic methods
