+++ pandas/core/indexes/extension.py
@@ -26,7 +26,6 @@ from pandas.util._exceptions import rewrite_exception
 
 from pandas.core.dtypes.common import (
     is_dtype_equal,
-    is_object_dtype,
     pandas_dtype,
 )
 from pandas.core.dtypes.generic import (
@@ -45,7 +44,6 @@ from pandas.core.arrays._mixins import NDArrayBackedExtensionArray
 from pandas.core.arrays.base import ExtensionArray
 from pandas.core.indexers import deprecate_ndim_indexing
 from pandas.core.indexes.base import Index
-from pandas.core.ops import get_op_result_name
 
 if TYPE_CHECKING:
 
@@ -174,55 +172,6 @@ def _make_wrapped_comparison_op(opname: str):
     return wrapper
 
 
-def _make_wrapped_arith_op(opname: str):
-    def method(self, other):
-        if (
-            isinstance(other, Index)
-            and is_object_dtype(other.dtype)
-            and type(other) is not Index
-        ):
-            # We return NotImplemented for object-dtype index *subclasses* so they have
-            # a chance to implement ops before we unwrap them.
-            # See https://github.com/pandas-dev/pandas/issues/31109
-            return NotImplemented
-
-        try:
-            meth = getattr(self._data, opname)
-        except AttributeError as err:
-            # e.g. Categorical, IntervalArray
-            cls = type(self).__name__
-            raise TypeError(
-                f"cannot perform {opname} with this index type: {cls}"
-            ) from err
-
-        result = meth(_maybe_unwrap_index(other))
-        return _wrap_arithmetic_op(self, other, result)
-
-    method.__name__ = opname
-    return method
-
-
-def _wrap_arithmetic_op(self, other, result):
-    if result is NotImplemented:
-        return NotImplemented
-
-    if isinstance(result, tuple):
-        # divmod, rdivmod
-        assert len(result) == 2
-        return (
-            _wrap_arithmetic_op(self, other, result[0]),
-            _wrap_arithmetic_op(self, other, result[1]),
-        )
-
-    if not isinstance(result, Index):
-        # Index.__new__ will choose appropriate subclass for dtype
-        result = Index(result)
-
-    res_name = get_op_result_name(self, other)
-    result.name = res_name
-    return result
-
-
 def _maybe_unwrap_index(obj):
     """
     If operating against another Index object, we need to unwrap the underlying
@@ -290,23 +239,6 @@ class ExtensionIndex(Index):
     __gt__ = _make_wrapped_comparison_op("__gt__")
     __le__ = _make_wrapped_comparison_op("__le__")
     __ge__ = _make_wrapped_comparison_op("__ge__")
-
-    __add__ = _make_wrapped_arith_op("__add__")
-    __sub__ = _make_wrapped_arith_op("__sub__")
-    __radd__ = _make_wrapped_arith_op("__radd__")
-    __rsub__ = _make_wrapped_arith_op("__rsub__")
-    __pow__ = _make_wrapped_arith_op("__pow__")
-    __rpow__ = _make_wrapped_arith_op("__rpow__")
-    __mul__ = _make_wrapped_arith_op("__mul__")
-    __rmul__ = _make_wrapped_arith_op("__rmul__")
-    __floordiv__ = _make_wrapped_arith_op("__floordiv__")
-    __rfloordiv__ = _make_wrapped_arith_op("__rfloordiv__")
-    __mod__ = _make_wrapped_arith_op("__mod__")
-    __rmod__ = _make_wrapped_arith_op("__rmod__")
-    __divmod__ = _make_wrapped_arith_op("__divmod__")
-    __rdivmod__ = _make_wrapped_arith_op("__rdivmod__")
-    __truediv__ = _make_wrapped_arith_op("__truediv__")
-    __rtruediv__ = _make_wrapped_arith_op("__rtruediv__")
 
     # ---------------------------------------------------------------------
     # NDarray-Like Methods
